requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.3.0
cachetools==5.3.1
celery[redis]==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.54.0
django-debug-toolbar==4.1.0
django-filter==23.1
django-jazzmin==2.6.0
django-redis==5.2.0
django-rest-multiple-models==2.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.5
ed25519==1.5
env==0.1.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.0
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
product==0.1.1.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.3
Token==1.8.6
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarg==0.1.9
